Implement Call for OpEthApi
