+++ crates/rpc/rpc-eth-api/src/helpers/call.rs
@@ -29,8 +29,6 @@ use reth_rpc_types::{
 };
 use revm::{Database, DatabaseCommit};
 use revm_inspectors::access_list::AccessListInspector;
-#[cfg(feature = "optimism")]
-use revm_primitives::OptimismFields;
 use tracing::trace;
 
 use super::{LoadBlock, LoadPendingBlock, LoadState, LoadTransaction, SpawnBlocking, Trace};
@@ -824,6 +822,8 @@ pub trait Call: LoadState + SpawnBlocking {
             )?;
 
         let gas_limit = gas.unwrap_or_else(|| block_env.gas_limit.min(U256::from(u64::MAX)).to());
+
+        #[allow(clippy::needless_update)]
         let env = TxEnv {
             gas_limit: gas_limit
                 .try_into()
@@ -843,8 +843,7 @@ pub trait Call: LoadState + SpawnBlocking {
             // EIP-7702 fields
             authorization_list: None,
             // authorization_list: TODO
-            #[cfg(feature = "optimism")]
-            optimism: OptimismFields { enveloped_tx: Some(Bytes::new()), ..Default::default() },
+            ..Default::default()
         };
 
         Ok(env)
